Move formatRelativeTime into timeFormatter
Keeps one module as the only place a time is formatted, so the lint rule
confining date-fns formatting has a single exemption rather than two. Its tests
move into timeFormatter.test.ts alongside the other formatters.
