+++ shared/CLAUDE.md
@@ -26,10 +26,9 @@ Guidance for working in `shared/`. See the [root CLAUDE.md](../CLAUDE.md) for th
 
 - **`result.ts`** — the `Result<T, E>` tagged union and its constructors (see below).
 - **`signupTimes.ts`** — lottery/direct sign-up window calculations (`getLotterySignupStartTime`/`…EndTime`, `getDirectSignupStartTime`/`…EndTime`, in-progress/ended predicates). Applies the parent start-time override (see below). The "open the previous evening at a fixed hour" path does its arithmetic on a `TZDate` in the event timezone and names the target hour explicitly, so it lands on the same wall-clock hour whether the previous day was 23, 24 or 25 hours long; `signupTimes.test.ts` pins both transitions. A `TZDate` serializes to its own offset rather than to UTC, so it is converted back to a plain `Date` before being returned — these values are stored and compared as ISO strings elsewhere.
-- **`timeFormatter.ts`** — date-fns display formatting; everything is forced to the `Europe/Helsinki` timezone via `@date-fns/tz`. The `getLocal*` variants deliberately use the viewer's own timezone instead, for text that contrasts local time with Finnish time.
+- **`timeFormatter.ts`** — date-fns display formatting; everything is forced to the `Europe/Helsinki` timezone via `@date-fns/tz`. The `getLocal*` variants deliberately use the viewer's own timezone instead, for text that contrasts local time with Finnish time. `formatRelativeTime` (date-fns `formatDistance`) lives here too, so this is the only module that formats a time; `timeFormatter.test.ts` pins every string in both languages, because the event log shows them to attendees.
 - **`timezone.ts`** — the `TIMEZONE` constant.
 - **`setLocale.ts`** — the active date-fns locale. date-fns takes a locale per call rather than from global state, so the current one is held here and read by the formatters. It is also a store (`subscribeToLocale`/`getLocaleSnapshot`): everything the formatters produce depends on this module state, which React cannot see, so `useTimeNow` subscribes and gives the instant a new identity on a language change - otherwise a memoized weekday keeps rendering in the language it was first read in. Locales are imported one at a time (`date-fns/locale/fi`), never from the `date-fns/locale` barrel, which pulls in all 724. The Finnish locale is customized: date-fns abbreviates weekdays as `torst.` where the UI has always shown `to`, and English needs the abbreviated width unchanged, so this cannot be solved with a different format token.
-- **`relativeTime.ts`** — `formatRelativeTime`, a thin wrapper over date-fns `formatDistance` that supplies the active locale. `time.test.ts` pins every string in both languages, because the event log shows them to attendees.
 - **`timeComparison.ts`** — `isSameOrAfter`/`isSameOrBefore` (date-fns has no direct equivalent) plus the two bounded-range checks whose edge behaviour matters.
 - **`isLotterySignupProgramItem.ts`** — predicate: does an item use two-phase (lottery) sign-up?
 - **`isDirectSignupAlwaysOpen.ts`** — predicate: is an item's direct sign-up always open? Combines the manual `directSignupAlwaysOpenIds` config list with a programmatic check for the `Tag.PRE_CONVENTION_WEEK` tag (pre-convention-week items always use direct sign-up, even lottery program types like RPGs). Consumed by `isLotterySignupProgramItem` and `getDirectSignupStartTime`.
@@ -78,6 +77,6 @@ When adding new code that writes `signedToStartTime`, follow this split. When ad
 ## Conventions
 
 - Use enums/`as const` for closed sets (program types, sign-up types, user groups); use the `Result` tagged union for fallible operations.
-- All time formatting must go through `timeFormatter.ts` with the `Europe/Helsinki` timezone — never rely on the host's local timezone. A `no-restricted-syntax` rule in the root `eslint.config.ts` enforces this by confining `.format()` to `timeFormatter.ts`, as a file-scoped override rather than an inline `eslint-disable` comment, so that file keeps the rule's other restrictions.
+- All time formatting must go through `timeFormatter.ts` with the `Europe/Helsinki` timezone — never rely on the host's local timezone. A `no-restricted-imports` rule in the root `eslint.config.ts` enforces this by confining date-fns's `format`/`formatDistance`/`formatRelative` to that one module, as a file-scoped override rather than an inline `eslint-disable` comment, so the file keeps the rule's other restrictions.
 - **Times are `Date`**, formatted and compared with date-fns; `dayjs` is gone and a `no-restricted-imports` rule keeps it out. The migration was motivated by types: `dayjs()` accepted `null`/`undefined`, so a nullable time compiled fine and produced an Invalid Date whose every comparison returned `false` — a wrong answer rather than an error. date-fns rejects nullish input at compile time, which surfaced several genuinely nullable values that had been silently absorbed.
 - No database migrations: events run on fresh, short-lived DBs, so enum/shape changes here need no migration or compatibility shim (see [root CLAUDE.md](../CLAUDE.md)).
